Aktualisierungen auf Neu_Oktober29

+++ Backend/src/main/java/com/rezepte_app/repository/TagRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface TagRepository extends JpaRepository<Tag, Integer> {
+public interface TagRepository extends JpaRepository<Tag, Long> {
 
-    Optional<Tag> findByLabelAndId(String label, int id);
+    Optional<Tag> findByLabelAndId(String label, Long id);
 }
